+++ changelog.md
@@ -12,6 +12,8 @@
 ### Enhancements
 * You can now add sites to the `TestNetworkBuilder` via `addSite`.
 * You can now start the `AssignToFeeder` trace from a specified `Terminal` rather than all feeder heads.
+* When processing feeder assignments, all LV feeders belonging to a dist substation site will now be considered energized when the site is energized by a
+  feeder.
 
 ### Fixes
 * None.